Add more classes to achieve more oop-like practices
