+++ data/data.txt
@@ -1,4 +1,5 @@
-T | 0 | borrow book
 T | 1 | another todo task
-E | 0 | project meeting | 2024-12-01 | 2024-12-02
-D | 0 | return books | 2024-09-02
+T | 0 | a a
+T | 1 | b b
+D | 0 | c c | 2024-01-10
+E | 0 | d d | 2024-01-01 | 2024-02-02
